feature/86cw4nf56: Created the individual service helper function
